+++ frontend/app/src/components/settings/StatisticsGraphSettings.vue
@@ -1,53 +1,19 @@
 <script setup lang="ts">
-const emit = defineEmits(['updated']);
-
-const multiplier = ref('0');
-const visible = ref(false);
-const numericMultiplier = computed(() => {
-  const multi = Number.parseInt(get(multiplier));
-  return isNaN(multi) ? 0 : multi;
-});
-const invalid = computed(() => {
-  const numericValue = Number.parseInt(get(multiplier));
-  return isNaN(numericValue) || numericValue < 0;
-});
-
-const { update } = useSettingsStore();
-const { ssfGraphMultiplier, balanceSaveFrequency } = storeToRefs(
-  useGeneralSettingsStore()
-);
-
 const { t } = useI18n();
 
-const updateSetting = async () => {
-  await update({
-    ssfGraphMultiplier: get(numericMultiplier)
-  });
-  emit('updated');
-  set(visible, false);
-};
+const emit = defineEmits<{
+  (e: 'updated'): void;
+}>();
 
-const multiplierSetting = computed(() => get(ssfGraphMultiplier).toString());
+const updated = () => emit('updated');
 
-const period = computed(() => {
-  const multi = get(numericMultiplier);
-  if (multi <= 0) {
-    return 0;
-  }
-  return multi * get(balanceSaveFrequency);
-});
-
-onMounted(() => {
-  set(multiplier, get(multiplierSetting));
-});
-
-watch(multiplierSetting, value => set(multiplier, value.toString()));
+const showMenu: Ref<boolean> = ref(false);
 </script>
 
 <template>
   <v-menu
-    v-model="visible"
-    max-width="300px"
+    v-model="showMenu"
+    max-width="500px"
     min-width="280px"
     left
     :close-on-content-click="false"
@@ -61,32 +27,16 @@ watch(multiplierSetting, value => set(multiplier, value.toString()));
         <v-icon>mdi-dots-vertical</v-icon>
       </menu-tooltip-button>
     </template>
-    <card>
-      <template #title>{{ t('statistics_graph_settings.title') }}</template>
-      <template #subtitle>
-        {{ t('statistics_graph_settings.subtitle') }}
-      </template>
-      <v-text-field
-        v-model="multiplier"
-        type="number"
-        outlined
-        :label="t('statistics_graph_settings.label')"
-      />
 
-      <span v-if="period === 0">{{ t('statistics_graph_settings.off') }}</span>
-      <span v-else>
-        {{ t('statistics_graph_settings.on', { period }) }}
-      </span>
+    <card>
+      <ssf-graph-multiplier-setting @updated="updated()" />
+      <v-divider class="my-4" />
+      <infer-zero-timed-balances-setting @updated="updated()" />
 
       <template #buttons>
         <v-spacer />
-        <v-btn
-          depressed
-          color="primary"
-          :disabled="invalid"
-          @click="updateSetting()"
-        >
-          {{ t('common.actions.save') }}
+        <v-btn depressed color="primary" @click="showMenu = false">
+          {{ t('common.actions.close') }}
         </v-btn>
       </template>
     </card>